Refactor plugins
